+ Implement file api for nspawn-client channel
Implement create_file/write_fd/close_fd functions over the nspawn-client API, to
work with "remote" files over the unix socket. In this way, we can open multiple
files at once, stream content to them independently, and control when they are closed.

This API should be pretty easy to extend more in the future, for even better
control over the nspawn client.

Additionally, the nspawn command has been improved, so it now uses a temporary
dir for both the socket and the client itself. After running the machine, the
empty mount dir is removed by raptor, to ensure a clean end result.