* Make raptor require privileges for "build" and "run" commands
We can't run `systemd-nspawn` in realistic scenarios without sudo, but
previously this was applied as `sudo` directly in front of `systemd-nspawn`,
internally by raptor.

However, this creates problems. For `build`, the output files (`layers/`), and
for `run` mode, the `state` dir, are still not writable by the unprivileged user.

Instead, check for privileges on startup, and write a helpful message when
bailing out.