* Major refactor of Loader vs RaptorBuilder
Make Loader use DashMap for sources, packages and programs, which
permits almost all operations to run on &Loader intstead of &mut Loader.

Move origin tracking out of Loader, and instead provide &mut Vec<Origin>
to load functions, which also makes it easier for call sites to synthesize
origin context. For example, this solves the complicated problem of having
good error messages for a file with an invalid FROM instruction.

With these changes, a large number of operations are now non-mut, leading
to a cleaner api, and better potential for running operations in parallel.