+ Implement custom parser for www-authenticate header
The http header www-authenticate has a good deal of complexity to it.

At the same time, we need to be able to parse this header, to support generic
and automatic token logins for container registries, without a hard-coded list
of certain known hosts.

So we implement a custom parser using the Logos crate. It's not (currently)
trying to get every single edge case from the RFC covered, but should be good
enough for general use with token auth for registries.

Unit tests are included for various real responses, as well as the RFC example.