* Rework dregistry crate, support multiple APIs, improve caching
Change dregistry internals, making it support both v1 and v2 APIs,
as well as manifest indices.

Also, instead of having a hard-coded table of registries that require
authentication, respond correctly to the `www-authenticate` header that is
returned from the initial attempt to fetch the resource.

This requires hooking in the new `authparse` module, and using that to construct
the authentication request.

This new code has been tested on a variety of images on 7 major registries
(Docker, Quay, Amazon, Azure, Gitlab, Github, Google).