* Rework cache key calculation for deterministic hashing
The cache key is now based on the hash of the flattened instruction set, and the
hash of contents of direct source files.

Direct sources are those mentioned in COPY, RENDER, etc.

However, instructions that don't directly affect builds, or only affect runs
(FROM, ENTRYPOINT, MOUNT, etc) are no longer included in the hash.